+++ sdk/dotnet/GetDevice.cs
@@ -111,6 +111,10 @@ namespace Pulumi.Tailscale
         /// </summary>
         public readonly string Id;
         public readonly string Name;
+        /// <summary>
+        /// The user associated with the device
+        /// </summary>
+        public readonly string User;
 
         [OutputConstructor]
         private GetDeviceResult(
@@ -118,11 +122,14 @@ namespace Pulumi.Tailscale
 
             string id,
 
-            string name)
+            string name,
+
+            string user)
         {
             Addresses = addresses;
             Id = id;
             Name = name;
+            User = user;
         }
     }
 }