+++ sdk/go/tailscale/init.go
@@ -22,6 +22,8 @@ func (m *module) Construct(ctx *pulumi.Context, name, typ, urn string) (r pulumi
 	switch typ {
 	case "tailscale:index/acl:Acl":
 		r = &Acl{}
+	case "tailscale:index/deviceAuthorization:DeviceAuthorization":
+		r = &DeviceAuthorization{}
 	case "tailscale:index/deviceSubnetRoutes:DeviceSubnetRoutes":
 		r = &DeviceSubnetRoutes{}
 	case "tailscale:index/dnsNameservers:DnsNameservers":
@@ -64,6 +66,11 @@ func init() {
 	pulumi.RegisterResourceModule(
 		"tailscale",
 		"index/acl",
+		&module{version},
+	)
+	pulumi.RegisterResourceModule(
+		"tailscale",
+		"index/deviceAuthorization",
 		&module{version},
 	)
 	pulumi.RegisterResourceModule(